Restructure, remove and comment other things
